requirements.txt
streamlit>=1.24.0
requests>=2.28.0
python-dotenv>=0.19.0
pandas>=1.5.0
plotly>=5.13.0
psutil>=5.9.0
pydantic